+++ packages/cqrs/lib/src/command_result.dart
@@ -16,10 +16,7 @@ import 'transport_types.dart';
 
 /// The result of running a [Command].
 class CommandResult {
-  const CommandResult(
-    this.errors, {
-    @Deprecated("Success is derived from the `errors` emptiness.") bool success,
-  }) : assert(errors != null);
+  const CommandResult(this.errors) : assert(errors != null);
 
   /// Creates a success [CommandResult] without any errors.
   const CommandResult.success() : errors = const [];
@@ -46,17 +43,21 @@ class CommandResult {
   /// Checks whether this [CommandResult] contains a provided error `code` in
   /// its validation errors.
   bool hasError(int code) => errors.any((error) => error.code == code);
+
+  /// Checks whether this [CommandResult] contains a provided error `code` in
+  /// its validation errors related to the `propertyName`.
+  bool hasErrorForProperty(int code, String propertyName) => errors
+      .any((error) => error.code == code && error.propertyName == propertyName);
 }
 
 /// A validation error.
 class ValidationError {
-  const ValidationError(this.code, this.message)
-      : assert(code != null),
-        assert(message != null);
+  const ValidationError(this.code, this.message, this.propertyName);
 
   ValidationError.fromJson(Map<String, dynamic> json)
       : code = json['ErrorCode'] as int,
-        message = json['ErrorMessage'] as String;
+        message = json['ErrorMessage'] as String,
+        propertyName = json['PropertyName'] as String;
 
   /// Code of the validation error.
   final int code;
@@ -64,6 +65,9 @@ class ValidationError {
   /// Message describing the validation error.
   final String message;
 
+  /// Path to the property which caused the error.
+  final String propertyName;
+
   @override
-  String toString() => '$code: $message';
+  String toString() => '[$propertyName] $code: $message';
 }