+++ packages/cqrs/test/command_result_test.dart
@@ -3,8 +3,8 @@ import 'package:test/test.dart';
 
 void main() {
   group('CommandResult', () {
-    const error1 = ValidationError(1, 'First error');
-    const error2 = ValidationError(2, 'Second error');
+    const error1 = ValidationError(1, 'First error', 'Property1');
+    const error2 = ValidationError(2, 'Second error', 'Property2');
 
     group('fields values are correct', () {
       test('when constructed without errors', () {
@@ -58,6 +58,31 @@ void main() {
       });
     });
 
+    group('hasErrorForProperty returns correct values', () {
+      test('when there are some errors present', () {
+        const result = CommandResult([error1, error2]);
+
+        expect(result.hasErrorForProperty(1, 'Property1'), true);
+        expect(result.hasErrorForProperty(2, 'Property2'), true);
+      });
+
+      test('when there are errors but for different properties', () {
+        const result = CommandResult([error1, error2]);
+
+        expect(result.hasErrorForProperty(1, 'Property2'), false);
+        expect(result.hasErrorForProperty(2, 'Property1'), false);
+      });
+
+      test('when there are no errors present', () {
+        const result = CommandResult([]);
+
+        expect(result.hasErrorForProperty(1, 'Property1'), false);
+        expect(result.hasErrorForProperty(2, 'Property1'), false);
+        expect(result.hasErrorForProperty(2, 'Property1'), false);
+        expect(result.hasErrorForProperty(2, 'Property2'), false);
+      });
+    });
+
     group('is correctly deserialized from JSON', () {
       test('with some validation errors', () {
         final result = CommandResult.fromJson({
@@ -66,6 +91,7 @@ void main() {
             {
               'ErrorCode': 12,
               'ErrorMessage': 'This is an error.',
+              'PropertyName': 'Property',
             },
           ],
         });
@@ -77,11 +103,11 @@ void main() {
               .having(
                 (r) => r.errors,
                 'errors',
-                contains(
-                  isA<ValidationError>()
-                      .having((e) => e.code, 'code', 12)
-                      .having((e) => e.message, 'message', 'This is an error.'),
-                ),
+                contains(isA<ValidationError>()
+                    .having((e) => e.code, 'code', 12)
+                    .having((e) => e.message, 'message', 'This is an error.')
+                    .having(
+                        (e) => e.propertyName, 'propertyNamme', 'Property')),
               ),
         );
       });